Add POC for double template evaluation
In the original template, the placeholders for positional arguments or
named variables (as per user's choice) are to be specified using the
'placeholder' function added to the template environment.

Double template evaluation means the input template will be evaluated
twice subsequently.

In the first step, all variables and function calls will be
evaluated. Variables are expected to be of type boolean with var names
prefixed with 'cond__' as a convention. The placeholder function calls
will get evaluated to jinja expressions. Example:

{{ placeholder('myvar') }} will get evaluated to '{{ myvar }}'

In other words, the output of first rendering is also a jinja
template.

This intermediate representation can now be rendered into a query that
can either be,

1. used in production code (with positional arguments or
   named variables) or,

2. used to generate pgTap tests (POC pending)

An example template is also added. The query is based on the Chinook
database[1].

[1]: https://github.com/lerocha/chinook-database