Derive positional arguments from jinja vars in intermediate query
Since the intermediate query will also contain jinja variables, we
need to treat it as a jinja template and render it again. However,
this time the variables need to be replaced with positional
arguments (postgres). This is done using minijinja's
'template.undeclared_variables' function and a bit of regex matching.