Add noformat markers
These can be used to indicate that a certain code block shouldn't be
formatted, instead rendered as it is. The main use case for this is to
preserve custom formatting in tests to aid their readability.

The rendered sql files can then be formatted by running pg_format as
follows,

  pg_format -M -p 'start(noformat).+end(noformat)' songs_ps_test.sql

and the formatting inside the 'noformat' block will be preserved.