Remove support for inline queries in test templates
The approach of creating a prepared statement before the test just
works, and I don't see any advantage of providing a way to include the
query inline. Rather it unnecessarily adds to the complexity.

Will reconsider the inline query approach if a use case is found
later.

The relevant keys in the manifest files have been removed along with
the 'songs_test.sql' file that doesn't use prepared statements.