Derive the optional output fields at the time of decoding itself
The 'output' fields in structs 'Query' and 'TestTemplate' were earlier
defined as options as they can be optionally specified by the user in
the manifest file. The idea was that the value could be derived from
other fields i.e 'query.id', and test_'template.path' when
required (at the time of writing the output to files). However, that
requires the 'queries_output_dir' and 'tests_output_dir' to be known
at the time of writing output to the files. While it can be done,
there is no advantage of the additional complexity.

This commit modifies the 'decode' methods on the above structs to
derive the 'output' field values if not specified. Hence the 'output'
fields are no longer defined as 'Option'.