Make the docker image more generic
So that the same image can be used to run tests against any
database. The steps for setting up the db schema is now moved out of
the Dockerfile and into the run-tests script.

This commit also moves the docker directory in project root as it's no
longer coupled with the chinook example.