Add 'chinook' example which is consistent with the documentation
When the 'getting started' section of documentation was written, it
deviated a bit from the existing 'examples/chinook' files in the
repo. Hence in an earlier commit, the existing example dir was renamed
to 'chinook_old'. This commit adds a new one which is consistent with
the documentation.

In a later commit, the 'chinook_old' directory would be deleted.