Make 'all_conds' and 'conds' fields optional in manifest
In the manifest entries for 'query_templates' and 'queries' the
'all_conds' and 'conds' fields respectively are now optional. When not
specified, the default value of empty HashSet will be considered.

This improves conciseness for entries that don't have any conditional
templating in the query templates. A legit use case for such entries
without any conditional templating is to be able to embed them in
pgTAP tests.