Exclude backup files when listing template and output files
Some editors, such as emacs, creates a backup file at the same
location with '~' suffix added to the name of the original file. We
need to list template and output files for various use cases such as
including files that may be missing in the manifest in the summary
output or emitting warnings about such files. In these cases, we can
exclude the backup files.