Implement configurable 'layouts'
This commit adds support for two kind of layouts:

1. one-file-one-query: each query will be written to a differnt file
2. one-file-all-queries: all queries will be written to the same file

'one-file-one-query' is the default if not specified, and it's the
same as the earlier behaviour. Hence this commit is backward
compatible.

This commit contains all the code to get the functionality working -
input validation, output generation (that's why it's a pretty big
commit). However but much of it can be improved or optimized. It will
be done in later commits.