Remove unnecessary lifetime specifier
The 'validate' methods on 'Query' and 'Queries' structs were earlier
implemented with two lifetime specifiers for two arguments. Due to
this, a lifetime bound also had to be specified. Turns out that this
is not required as the compiler is able to check with a single
specifier for both arguments.