Configure name_tagger in chinook example
Because when a new project is created, name_tagger is configured with
style = kebab-case. Since the chinook example is used in the 'getting
started' section of the docs, it makes sense to configure the name
tagger in the actual example similarly.

Also update the getting started section in the docs.