Use 'OnceCell' to initialize the arg vector for PgFormatter lazily
The args are passed to the command when running it, and hence are only
required for formatting sql. So the 'OnceCell' is initialized when the
'format' method is called for the first time. Sort of like a cache but
implemented using internal mutability instead of an Option (in which
case, we'd need a unique/mutable reference to call the 'format'
method.

As 'OnceCell' was added to the standard library only in version
1.70.0, it's specified as the 'rust-version' (MSRV) in Cargo.toml
file.