Add 'ExternalFormatter' trait
This trait provides methods for formatting sql by calling externally
installed formatting programs such as 'pg_format', 'sqlfluff' (not
supported yet) etc. The trait replaces the earlier
'formatters::util::Cmd' abstraction for running external commands.

This is done to provide an easy and well defined way to add
support for external formatters.

The older code in 'formatters::util' is marked as 'dead_code'. It will
be removed in a later commit.