Add an abstraction for serializing simple toml tables with comments
The problem with using the 'serde' crate with TOML is that while
deserialization handles comments, there is no way to add comments in
the TOML file while serializing. Essentially, the
deserialization-serialization process is lossy in nature.

This commit introduces an abstraction using which key value pairs
along with explicit comments can be added while serialization.

This abstraction is meant to be used for adding formatter config in
the TOML file that's generated by the scaffolding mechanism ('init'
command). That change will be done in a separate commit.