Add a 'Configurable' trait for formatters
This trait has one required method 'to_toml_table' that returns a
'SerializableTomlTable' that can be used to serialize the formatter
config in the toml manifest file.

It is also added as a supertrait of 'ExternalFormatter'. This means
that all external formatters must also implement this trait.

Right now, this trait is only meant for formatters. Hence it's added
under the top level 'formatters' module. But if it turns out that it
has a wider use case, we can move it to the 'toml' module in future.