Allow users to select a formatter during project initialization
Users also have the option to opt out of formatting by selecting the
'None' option.

This commit builds on top of previous commit that added code for
discovering the formatters installed on user's system.

'inquire' crate has been introduced as a new dependency for displaying
the formatter selection prompt to the user.