Move code for writing config file to formatters module
As part of scaffolding/project initialization, a config file is
generated for the formatter if applicable. Previously this code was in
the scaffolding module. However, as support for multiple formatters
has been added in recent commits, it makes sense for the respective
formatter module to provide this functionality.

To facilitate that, two new methods 'config_file' (required) and
'generate_config_file' (provided) are added to the 'Configurable'
trait.