Refactor name tagging code for reusability
Added a separate method to generate a name tag comment line. This can
be used in other places e.g. to check if a line is a name tag comment
when parsing the output SQL file when layout =
one-file-all-queries (this will be implemented in future commits to
get the 'status' command working with 'one-file-all-queries' layout).