Fix 'status' command to work with 'one-file-all-queries' layout
The status command compares the rendered SQL with contents of the
existing output query files. This is straightforward for layout =
'one-file-one-query'. But in case of 'one-file-all-queries', the
output showed 'modified' for all. In other words, it was a bug, since
the layout functionality was introduced.

This commit fixes it parsing the combined SQL file and creating an
index of query id mapping to the rendered SQL. An abstraction
'QueryOutputReader' is provided that can transparently 'read' SQL
either from actual individual files (layout='one-file-one-query') or
from the above index (layout='one-file-all-queries').

This commit employs certain workarounds for dealing with whitespace
differences due, which are documented inline.