Fix 'status' command when formatting is disabled
When formatting is disabled, there's a chance that a rendered query
doesn't have a trailing newline. This causes a problem when comparing
the rendered SQL output for an individual query with the SQL parsed
from a combined SQL output file (which has a newline after every SQL,
as added by the render command). This results in a false-positive
'modified' status in the 'status' command output.

To fix this,

1. we're ensuring that during comparision, the rendered SQL
   contains a trailing newline.

2. the combined SQL output file always ends in a newline